Move plugin framework provider and test utility files to packages (#8226)
* Move plugin framework provider and test utility files to the specific package

* Don't replace Nprintf

* Fix the rebasing error

* Fix tgc
