+++ google/data_source_dns_record_set_test.go
@@ -89,7 +89,7 @@ func testAccCheckDnsRecordSetDestroyProducerFramework(t *testing.T) func(s *terr
 				continue
 			}
 
-			p := GetFwTestProvider(t)
+			p := acctest.GetFwTestProvider(t)
 
 			url, err := acctest.ReplaceVarsForFrameworkTest(&p.FrameworkProvider.FrameworkProviderConfig, rs, "{{DNSBasePath}}projects/{{project}}/managedZones/{{managed_zone}}/rrsets/{{name}}/{{type}}")
 			if err != nil {
