+++ google/framework_provider_test.go
@@ -3,38 +3,13 @@
 package google
 
 import (
-	"context"
 	"fmt"
-	"io/ioutil"
-	"log"
 	"regexp"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-framework/diag"
-	"github.com/hashicorp/terraform-plugin-framework/provider"
-	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
-	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-provider-google/google/acctest"
 )
-
-func GetFwTestProvider(t *testing.T) *frameworkTestProvider {
-	configsLock.RLock()
-	fwProvider, ok := fwProviders[t.Name()]
-	configsLock.RUnlock()
-	if ok {
-		return fwProvider
-	}
-
-	var diags diag.Diagnostics
-	p := NewFrameworkTestProvider(t.Name())
-	configureApiClient(context.Background(), &p.FrameworkProvider, &diags)
-	if diags.HasError() {
-		log.Fatalf("%d errors when configuring test provider client: first is %s", diags.ErrorsCount(), diags.Errors()[0].Detail())
-	}
-
-	return p
-}
 
 func TestAccFrameworkProviderMeta_setModuleName(t *testing.T) {
 	// TODO: https://github.com/hashicorp/terraform-provider-google/issues/14158
@@ -54,76 +29,6 @@ func TestAccFrameworkProviderMeta_setModuleName(t *testing.T) {
 			},
 		},
 	})
-}
-
-func TestFrameworkProvider_impl(t *testing.T) {
-	var _ provider.ProviderWithMetaSchema = New("test")
-}
-
-func TestFrameworkProvider_CredentialsValidator(t *testing.T) {
-	cases := map[string]struct {
-		ConfigValue          func(t *testing.T) types.String
-		ExpectedWarningCount int
-		ExpectedErrorCount   int
-	}{
-		"configuring credentials as a path to a credentials JSON file is valid": {
-			ConfigValue: func(t *testing.T) types.String {
-				return types.StringValue(testFakeCredentialsPath) // Path to a test fixture
-			},
-		},
-		"configuring credentials as a path to a non-existant file is NOT valid": {
-			ConfigValue: func(t *testing.T) types.String {
-				return types.StringValue("./this/path/doesnt/exist.json") // Doesn't exist
-			},
-			ExpectedErrorCount: 1,
-		},
-		"configuring credentials as a credentials JSON string is valid": {
-			ConfigValue: func(t *testing.T) types.String {
-				contents, err := ioutil.ReadFile(testFakeCredentialsPath)
-				if err != nil {
-					t.Fatalf("Unexpected error: %s", err)
-				}
-				stringContents := string(contents)
-				return types.StringValue(stringContents)
-			},
-		},
-		"configuring credentials as an empty string is valid": {
-			ConfigValue: func(t *testing.T) types.String {
-				return types.StringValue("")
-			},
-		},
-		"leaving credentials unconfigured is valid": {
-			ConfigValue: func(t *testing.T) types.String {
-				return types.StringNull()
-			},
-		},
-	}
-
-	for tn, tc := range cases {
-		t.Run(tn, func(t *testing.T) {
-			// Arrange
-			req := validator.StringRequest{
-				ConfigValue: tc.ConfigValue(t),
-			}
-
-			resp := validator.StringResponse{
-				Diagnostics: diag.Diagnostics{},
-			}
-
-			cv := CredentialsValidator()
-
-			// Act
-			cv.ValidateString(context.Background(), req, &resp)
-
-			// Assert
-			if resp.Diagnostics.WarningsCount() > tc.ExpectedWarningCount {
-				t.Errorf("Expected %d warnings, got %d", tc.ExpectedWarningCount, resp.Diagnostics.WarningsCount())
-			}
-			if resp.Diagnostics.ErrorsCount() > tc.ExpectedErrorCount {
-				t.Errorf("Expected %d errors, got %d", tc.ExpectedErrorCount, resp.Diagnostics.ErrorsCount())
-			}
-		})
-	}
 }
 
 func TestAccFrameworkProviderBasePath_setInvalidBasePath(t *testing.T) {
