+++ google/fwprovider/framework_validators.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package fwprovider
 
 import (
 	"context"
