+++ google/resource_certificatemanager_certificate_upgrade_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-provider-google/google/acctest"
 )
 
 // Tests schema version migration by creating a certificate with an old version of the provider (4.59.0)
@@ -24,7 +25,7 @@ func TestAccCertificateManagerCertificate_migration(t *testing.T) {
 		},
 	}
 	newVersion := map[string]func() (*schema.Provider, error){
-		"mynewprovider": func() (*schema.Provider, error) { return TestAccProviders["google"], nil },
+		"mynewprovider": func() (*schema.Provider, error) { return acctest.TestAccProviders["google"], nil },
 	}
 
 	VcrTest(t, resource.TestCase{
