+++ google/resource_cloudbuild_trigger_upgrade_test.go
@@ -25,7 +25,7 @@ func TestAccCloudBuildTrigger_migration(t *testing.T) {
 		},
 	}
 	newVersion := map[string]func() (*schema.Provider, error){
-		"mynewprovider": func() (*schema.Provider, error) { return TestAccProviders["google"], nil },
+		"mynewprovider": func() (*schema.Provider, error) { return acctest.TestAccProviders["google"], nil },
 	}
 
 	VcrTest(t, resource.TestCase{
